correction

+++ ft_strncmp.c
@@ -23,5 +23,5 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	}
 	if (n == 0)
 		return (0);
-	return (s1[i] - s2[i]);
+	return ((int)(s1[i] - s2[i]));
 }
